fix API and tests for chainFunctions
un-declare global `process`
fixes #158

previous commit:
fixes #159
