+++ docs/core/dev-mode.md
@@ -12,16 +12,29 @@ Sets the development flag on or off.
 Sets the development flag according the the value of `process.env.NODE_ENV` once, at module load time.
 
 Also, sets the value of `process.env.NODE_ENV` whenever the development flag changes its value.
-so It's safe to use `if(process.env.NODE_ENV === 'production')`
+so It's safe to use `if(process.env.NODE_ENV === 'production')`. 
+This is designed to allow popular [uglify-envify `process.env` purging](https://github.com/hughsk/envify#purging-processenv) tecniques.
 
 ```ts
 import {devMode, setGlobalConfig} from 'wix-react-tools';
 
-setGlobalConfig(devMode.ON);
-// now process.env.NODE_ENV === 'development'
+// This is for passing typescript validation
+declare const process: {env : {[k:string]: any}};
 
+setGlobalConfig(devMode.ON);
+
+if (process.env.NODE_ENV === 'development'){
+    // this block will execute
+} else {
+    // this block will not execute
+}
 
 setGlobalConfig(devMode.OFF);
-// now process.env.NODE_ENV === 'production'
+
+if (process.env.NODE_ENV === 'production'){
+    // this block will execute
+} else {
+    // this block will not execute
+}
 
 ```
