+++ package.json
@@ -43,6 +43,7 @@
     "@types/chai-dom": "^0.0.5",
     "@types/lodash": "^4.14.69",
     "@types/mocha": "^2.2.41",
+    "@types/node": "^8.0.31",
     "@types/react": "^15.6.0",
     "@types/react-dom": "^15.5.3",
     "@types/sinon": "^2.3.2",
