+++ src/core/dev-mode.ts
@@ -1,12 +1,8 @@
 import {onGlobalConfig, setGlobalConfig} from "./config";
 
-declare global {
-    const process: any;
-}
-
 declare const global: any;
 declare const window: any;
-declare let process: any;
+declare let process: {env : {[k:string]: any}};
 
 const globalCtx = (typeof self === 'object' && self.self === self && self) ||
     (typeof global === 'object' && global['global'] === global && global) || window;
