+++ src/core/functional.ts
@@ -7,10 +7,11 @@ function chain<T extends Function>(first: T, last: T): T {
     } as any as T;
 }
 
-export const chainFunctions =  chain as typeof chain & {cached : typeof chain};
+export type Chain = <T extends Function>(first: T, last: T) => T;
+export const chainFunctions =  chain as Chain & {cached : Chain};
 chainFunctions.cached = memoize(chain);
 
 export const cachedChainFunctions = function chained(this: any, ...args: any[]) {
     console.warn(`cachedChainFunctions() is deprecated, use chainFunctions.cached() instead`);
-    chainFunctions.cached.apply(this, args);
-} as typeof chain;
+    return chainFunctions.cached.apply(this, args);
+} as Chain;
